requirements.txt
streamlit
pandas
matplotlib
seaborn
google-genai
python-dotenv
pillow
requests